취향분석 페이지 연결

+++ MainProject/src/main/java/music/is/mylife/dao/ListMapper.java
@@ -8,10 +8,10 @@ import music.is.mylife.vo.Song;
 
 public interface ListMapper {
 	//해당 유저의 모든 리스트 좋아요 수 
-	public int selectListLikesByUser(String user_id);
+	public double selectListLikesByUser(String user_id);
 	
 	//해당 유저의 모든 리스트에 지금까지 달린 모든 댓글 수
-	public int selectListCommentsNumByUser(String user_id);
+	public double selectListCommentsNumByUser(String user_id);
 	
 	public double selectListCountByUser(String user_id);
 	//리스트 페이지_화원
